remove extra space

+++ common/chapterTopics/deleteChapterTopic.js
@@ -23,7 +23,7 @@ module.exports = require('express').Router().post('/',async(req,res) =>
         }
         topicId = topic[0].id
         let checkUsed = await db.checkChapterTopicUsed(topicId)
-        if(checkUsed[0].isExist == 0){   
+        if(checkUsed[0].isExist == 0 && checkUsed[0].statusExist == 0){   
         // console.log(checkUsed,checkUsed[0])
             let deleteTopic = await db.deleteChapterTopic(uuid)
